test: Set up cypress (initial)
Very initial setup for running cypress for erm--most of the files stolen directly from `stripes-testing`.
We're branching out and doing this internally, which means we'll be succeptible to breaking changes in other apps. However in trade for this we get autonomy over our own testing repository, the ability to control our own ERM interactors etc without worrying about breaking other apps' tests, and the ability to use any special interactors we set up within stripes-erm-testing itself.

As mentioned above, this commit contains mostly copied config etc from stripes-testing to get this up and running. It can and will be tweaked moving forwards to help us write tests for ERM in the most streamlined way for the Bienenvolk team.

This also allows the likes of OA and Serials management to have full control over their e2e tests even when they are not yet accepted for flower release.

ERM-2626