fix: testIsolation false
testIsolation was causing the local storage and cookies to be cleared between each "it". This caused instant logout in stripes-core, so implement testIsolation: false to combat this. This reflects the config in stripes-testing.

In theory this should still isolate between "describe" but I have not tested this.

Updated version of cypress
Moved some before/afters around in locale-settings.cy.js
Added github action to potentially run cypress tests on a schedule (won't currently do anything with those though)

refs ERM-3280