test: fix cypress failures (#70)
* test: WORK IN PROGRESS DO NOT MERGE AS IS

Added some waits to solve flakiness (THIS IS NOT A VALID SOLUTION) and stumbled across what may be a legitimate problem with entitlement perms

* test: Fix... we were looking for an agreement line that had previously been deleted -.-

This remains a work in progress... not sure why the waits are now necessary. Guessing it's similar to how jest tests are needing waitFor() when using interactors. This isn't really tenable as a long term solution

Potentially have to reach out to stripes testing peeps for some information

* chore: Clean up package json test scripts

* ci: Run against snapshot not cypress test env (it's slow and flaky)

* test: Cypress refactors

Trying to tie up cypress test and make it WAY less flaky. Introduced a single wait to handle circumstance where "save without closing" wasn't waiting for save to actually finish before it moved onto the next test, which in turn threw everything out of whack

* test: Cleanup of test logic

Cleaned up "touch" logic and changed a wait to actually await a page interaction

Linting