test: Suppress certain errors (#90)
findDOMNode errors specifically cause us a lot of issues and come exclusively (thus far) from outside our module.

This is also true of props of type `on...` passed directly to html nodes. In general these are happening in Stripes not in our modules.

To that end, we wish to lower noise and allow ourselves to see actual errors and warnings in tests, and so we intercept console error itself and only display those errors not matching our list

This change lowers the number of errors in ui-agreements from 33 down to 4