feat: Icon interactor (#100)
Icons should now not instantly crash when in tests. This allows for testing of icons themselves.

Icon and IconElement interactors in place to help with this.
However, that same fix ends up causing a console proptypes error, in jest harness we want to very specifically quash that error but no others.